Ballot display conditions (check description for more info)
The ballot will display an unfilled ballot for a user if they have not voted and the election isn't over.
If the user has voted and the election is not over, then they will see their choices and will be unable to resubmit.
If the user has voted and the election is over then they will see their choices and the results of the election.